refactor: moved cs and js to seperate files, excluding root index.js and css
