+++ src/css/components/LikertScaleComponent.css
