+++ src/js/__tests__/Login.test.js
@@ -5,11 +5,11 @@
 import React, { act } from "react";
 import { render, screen, fireEvent, waitFor } from "@testing-library/react";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
-import MyProfile from "../pages/MyProfile";
-import Login from "../pages/Login";
+import Login from "../../pages/Login";
+import MyProfile from "../../pages/MyProfile";
 import mongoose from "mongoose";
 import bcrypt from "bcryptjs";
-import { getProfile } from "../services/ProfileService";
+import { getProfile } from "../../services/ProfileService";
 
 const localStorageMock = (function () {
   let store = {};
@@ -32,19 +32,18 @@ Object.defineProperty(window, "localStorage", {
 });
 
 jest.mock("../services/ProfileService", () => ({
+  registerUser: jest.fn(),
+  loginUser: jest.fn(),
   getProfile: jest.fn(),
-  logout: jest.fn(),
-  updateProfile: jest.fn(),
-  deleteProfile: jest.fn(),
 }));
 
-describe("MyProfile", () => {
+describe("Login", () => {
   const setup = () => {
     return render(
-      <MemoryRouter initialEntries={["/my-profile"]}>
+      <MemoryRouter initialEntries={["/login"]}>
         <Routes>
-          <Route path="/my-profile" element={<MyProfile />} />
           <Route path="/login" element={<Login />} />
+          <Route path="/my-profile" element={<MyProfile />} />
         </Routes>
       </MemoryRouter>
     );
@@ -86,27 +85,14 @@ describe("MyProfile", () => {
     localStorage.clear();
   });
 
-  test("renders profile details correctly", async () => {
-    getProfile.mockResolvedValueOnce({
-      name: "Test User",
-      profilePicture: {
-        data: "",
-        contentType: "image/png",
-      },
-      analytics: {
-        sessions: [],
-      },
-    });
-
-    localStorageMock.setItem("token", "fake-jwt-token");
+  test("renders login form correctly", () => {
     setup();
-
-    await waitFor(() => {
-      expect(screen.getByText("Welcome! Test User")).toBeInTheDocument();
-    });
+    expect(screen.getByLabelText("Email address")).toBeInTheDocument();
+    expect(screen.getByLabelText("Password")).toBeInTheDocument();
+    expect(screen.getByRole("button", { name: "Login" })).toBeInTheDocument();
   });
 
-  test("allows editing profile", async () => {
+  test("logs in a user and loads profile", async () => {
     getProfile.mockResolvedValueOnce({
       name: "Test User",
       profilePicture: {
@@ -118,18 +104,31 @@ describe("MyProfile", () => {
       },
     });
 
-    localStorageMock.setItem("token", "fake-jwt-token");
     setup();
-
-    fireEvent.click(await screen.findByText("Edit"));
-    expect(screen.getByLabelText("Name")).toBeInTheDocument();
-    fireEvent.change(screen.getByLabelText("Name"), {
-      target: { value: "Updated User" },
+    localStorageMock.setItem("token", "fake-jwt-token");
+    fireEvent.change(screen.getByLabelText("Email address"), {
+      target: { value: "test@example.com" },
     });
-    fireEvent.click(screen.getByText("Save"));
+    fireEvent.change(screen.getByLabelText("Password"), {
+      target: { value: "password123" },
+    });
+
+    await act(async () => {
+      fireEvent.click(screen.getByRole("button", { name: "Login" }));
+    });
+
+    // Check if the token is set in localStorage
+    expect(localStorageMock.getItem("token")).toBe("fake-jwt-token");
 
     await waitFor(() => {
-      expect(screen.getByText("Welcome! Updated User")).toBeInTheDocument();
+      expect(
+        screen.getByText((content, element) => {
+          return (
+            element.tagName.toLowerCase() === "h1" &&
+            content.includes("Welcome! Test User")
+          );
+        })
+      ).toBeInTheDocument();
     });
   });
 });
