+++ src/js/__tests__/Register.test.js
@@ -1,10 +1,10 @@
 import React, { act } from "react";
 import { render, screen, fireEvent } from "@testing-library/react";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
-import Register from "../pages/Register";
-import Login from "../pages/Login";
+import Register from "../../pages/Register";
+import Login from "../../pages/Login";
 import mongoose from "mongoose";
-import { registerUser } from "../services/ProfileService";
+import { registerUser } from "../../services/ProfileService";
 
 jest.mock("../services/ProfileService", () => ({
   registerUser: jest.fn(),
