+++ src/js/components/TextAnswerComponent.js
